+++ src/test/java/com/zxcmc/exort/api/ExortApi.java
@@ -1,0 +1,7 @@
+package com.zxcmc.exort.api;
+
+import org.bukkit.block.Block;
+
+public interface ExortApi {
+  boolean isExortChorusCarrier(Block block);
+}